requirement.txt
fastapi 
uvicorn 
pydantic 
sqlalchemy 
passlib bcrypt 
python-jose 
email-validator
python-multipart